+++ src/pages/components/features.js
@@ -1,0 +1,72 @@
+const features = [
+    {
+      title: <>Hey there Fellow Jitster!</>,
+      imageUrl: "img/undraw_code_review.svg",
+      description: (
+        <>
+          Jitsi Meet is a set of Open Source projects which empower users to
+          deploy secure, scalable and easy to use video conferencing platforms
+          with state-of-the-art video quality and features.
+  
+          <br/><br/>
+  
+          On This site you'll find documentation for all your Jitsi needs.
+  
+          <br /><br/>
+  
+          Get started <a href="docs/intro">here</a>.
+        </>
+      ),
+    },
+    {
+      title: <>Batteries included</>,
+      imageUrl: "img/undraw_real_time_sync.svg",
+      description: (
+        <>
+          Jitsi Meet supports all common browsers and also mobile devices.
+          
+          <br /><br />
+  
+          Our APIs allow developers to easily integrate Jitsi Meet into existing
+          applications, whether those are web based or native mobile apps.
+          
+          <br /><br />
+          
+          You can use our freely available instance at{" "}
+          <a href="https://meet.jit.si" target="_blank">
+            meet.jit.si
+          </a>{" "}
+          or self-host it yourself using our readily available Debian packages or
+          comprehensive Docker setup.
+        </>
+      ),
+    },
+    {
+      title: <>JaaS: Jitsi as a Service</>,
+      imageUrl: "img/undraw_going_up.svg",
+      description: (
+        <>
+          Looking for configuration flexibility without the complexity of
+          self-hosting and scalability management?
+          
+          <br /><br />
+          
+          Look no further than JaaS. 8x8 Jitsi as a Service (JaaS) is
+          an enterprise-ready video meeting platform that allows developers,
+          organizations and businesses to easily build and deploy video
+          solutions. With Jitsi as a Service we now give you all the power of
+          Jitsi running on our global platform so you can focus on building secure
+          and branded video experiences.
+          
+          <br /><br />
+          Check JaaS out{" "}
+          <a href="https://jaas.8x8.vc" target="_blank">
+            here
+          </a>
+          .
+        </>
+      ),
+    },
+  ];
+
+  export default features ;
